Finished implementing the screens for changing values of an OTP.

+++ otpimpl/hotp.h
@@ -19,7 +19,7 @@ private:
     std::string calculateSha1Hotp(unsigned char *key, size_t keyLength, uint64_t counter, size_t digits, bool addChecksum, int truncationOffset);
     std::string calculateHotpFromHmac(const unsigned char *hmac, size_t hmacSize, size_t digits, bool addChecksum, int truncationOffset);
     int64_t calcChecksum(int64_t otp, size_t digits);
-    unsigned char *dynamicTruncate(const unsigned char *hmac, size_t hmacSize, int truncateOffset);
+    unsigned char *dynamicTruncate(const unsigned char *hmac, size_t hmacSize, size_t truncateOffset);
 
     Hmac *mHmacToUse;
     bool mShouldDelete;
